Koshucode.Baala.Base.Relmap.Relmap

Description

Data structures for relation-to-relation mappings

Data

data Relmap v

Relation-to-relation mapping. A Relmap is correspond to a use of relational operator, that is pair of operator and its operand.

Constructors

RelmapSource HalfRelmap String [String]
Relmap that retrieves a relation from a dataset

RelmapConst HalfRelmap String (Rel v)
Relmap that has a constant relation

RelmapAlias HalfRelmap (Relmap v)
Relmap that has equavalent relmap

RelmapCalc HalfRelmap String (RelmapSub v) [Relmap v]
Relmap that maps relations to a relation

RelmapAppend (Relmap v) (Relmap v)
Relmap that joints two relmaps

RelmapName HalfRelmap String
Relmap reference

Function type

type RelmapSub v = [Rel v] -> Rel v -> Rel v

Function of relmap

type RelmapBin v = Relmap v -> Relmap v

Confluent operator, like binary operator

Append relmaps

This picture represents calculation of mapping input relation to output relation.

 input -[ relmap ]- output

Relmap A maps relation R1 to relation R2. Another relmap B maps R2 to R3.

 R1 -[ A ]- R2
 R2 -[ B ]- R3

Two relmaps A and B are jointed with intermidiate relation R2.

 R1 -[ A ]- R2 -[ B ]- R3

Or, we can draw a directly jointed picture. This whole structure is also RelmapAppend A B.

 R1 -[ A ]--[ B ]- R3

Selector

relmapSourceList :: Relmap v -> [Relmap v]

List of RelmapSource

relmapAppendList :: Relmap v -> [Relmap v]

Expand RelmapAppend to list of Relmap

Constructor

data OpUse v

Constructors

OpUse

Fields

opHalf :: HalfRelmap

opSub :: [Relmap v]

relmapSource :: OpUse v -> String -> [String] -> Relmap v

Retrieve relation from dataset

relmapCalc :: OpUse v -> String -> RelmapSub v -> Relmap v

Make a non-confluent relmap

relmapConfl :: OpUse v -> String -> RelmapSub v -> [Relmap v] -> Relmap v

Make a confluent relmap