Koshucode.Baala.Base.Struct.Half.HalfRelmap

Relmap implementation

makeConsRelmap

Arguments

:: [RelmapImplement v]
Implementations of relmap operator

-> ConsRelmap v
Relmap constructors

Make half and full relmap constructors.

data ConsRelmap v

Half and full relmap constructor

Constructors

ConsRelmap

Fields

consHalf :: RelmapHalfCons

consFull :: RelmapWholeCons v

data RelmapImplement v

Implementation of relmap operator. It consists of (1) operator name, (2) operand parser, (3) full constructor, and (4) usage of operator.

Constructors

RelmapImplement String OperandParser (RelmapFullCons v) [String]

type OperandParser = [TokenTree] -> [Named [TokenTree]]

Parser for operand of relational operator. This parsers docompose operand trees, and give a name to suboperand.

Half constructor

type RelmapHalfCons

Arguments

= String
Operator name

-> [SourceLine]

-> [TokenTree]
Operand as source trees

-> HalfRelmap
Result half relmap with empty subs

First step of relmap construction, make HalfRelmap from use of relational operator.

consHalfRelmap :: RelmapHalfCons -> [SourceLine] -> [TokenTree] -> HalfRelmap

Construct half relmaps from source trees.

Full constructor

type RelmapWholeCons v

Arguments

= HalfRelmap

-> AbortOr (Relmap v)
Result relmap

Second step of constructing relmap, make Relmap from contents of HalfRelmap.

type RelmapFullCons v

Arguments

= [Relmap v]
Subrelmaps computed by consFullRelmap

-> HalfRelmap

-> AbortOr (Relmap v)
Result relmap