Koshucode.Baala.Base.Struct.SectionIO

Reading section

sectionRead

Arguments

:: StringValue v

=> Section v
Section that is same type to result section

-> String
Source text

-> AbortOr (Section v)
Result section from source text

sectionFile :: StringValue v => Section v -> FilePath -> IO (AbortOr (Section v))

Read section from file.

Running section

runSectionJudges

Arguments

:: Value v

=> Section v
Editing section

-> [Judge v]
Input judges

-> AbortOr [Judge v]
Output judges

Convert input judges into output judges by editing section. This function calls runAssert for list of assertions in section, and runAssert calls runRelmap for relmap in assertion.

runSectionString :: Value v => Section v -> String -> AbortOr String

Convert input data into output data by editing section.

runSectionIO :: Value v => Section v -> IO ()

Run editing section.