Koshucode.Baala.Base.Syntax.Tree

Data type

data Tree a

Tree of leaf and branch

Constructors

TreeL a
Leaf. Terminal of tree.

TreeB Int [Tree a]
Branch. Paren-type and subtrees.

Parsing

tree :: Show a => ParenType a -> [a] -> Tree a

Convert a list of elements to a single tree.

trees :: Show a => ParenType a -> [a] -> [Tree a]

Convert a list of elements to trees.

untree :: TypeParen a -> Tree a -> [a]

Convert tree to list of tokens.

untrees :: TypeParen a -> [Tree a] -> [a]

Convert tree to list of tokens.

undouble :: Map (Tree a)

Simplify tree by removing double parens, like ((a)) to (a).

>>> undouble $ TreeB 0 [Tree 0 [TreeL "A", TreeL "B"]]
Tree 0 [TreeL "A", TreeL "B"]

Paren table

type ParenType a = a -> Int

Get a paren type.

type TypeParen a = Int -> (a, a)

Get parens from a type.

parenTable

Arguments

:: Eq a

=> [(Int, a, a)]
List of (type, opne, close)

-> (ParenType a, TypeParen a)
Paren functions.

Make ParenType and TypeParen functions from a type-open-close table.

Make paren/type functions from () and [].

>>> let (pt, tp) = parenTable [(1,'(',')'), (2,'[',']')]

Get paren types for each chars. Types of open parens are positive integer, and closes are negative.

>>> map pt "ab(cd[ef])g"
[0,0,1,0,0,2,0,0,-2,-1,0]

Get an open-close pair of parens from its type.

>>> tp 2
('[', ']')