Koshucode.Baala.Base.Relmap.Relmap

Description

Data structures for relation-to-relation mappings

Data

data Relmap v

Relation-to-relation mapping. A Relmap is correspond to a use of relational operator, that is pair of operator and its operand.

Constructors

RelmapSource HalfRelmap String [String]
Relmap that retrieves a relation from a dataset

RelmapConst HalfRelmap String (Rel v)
Relmap that has a constant relation

RelmapAlias HalfRelmap (Relmap v)
Relmap that has equavalent relmap

RelmapCalc HalfRelmap String (RelmapSub v) [Relmap v]
Relmap that maps relations to a relation

RelmapAppend (Relmap v) (Relmap v)
Relmap that joints two relmaps

RelmapName HalfRelmap String
Relmap reference

type RelmapSub v

Arguments

= [Rel v]
Subrelmaps

-> Map (Rel v)
Mapping relation to relation

Function of relmap

Append relmaps

This picture represents calculation of mapping input relation to output relation.

 input -[ relmap ]- output

Relmap A maps relation R1 to relation R2. Another relmap B maps R2 to R3.

 R1 -[ A ]- R2
 R2 -[ B ]- R3

Two relmaps A and B are jointed with intermidiate relation R2.

 R1 -[ A ]- R2 -[ B ]- R3

Or, we can draw a directly jointed picture. This whole structure is also RelmapAppend A B.

 R1 -[ A ]--[ B ]- R3

Selectors

relmapSourceList :: Relmap v -> [Relmap v]

List of RelmapSource

relmapAppendList :: Relmap v -> [Relmap v]

Expand RelmapAppend to list of Relmap

Use of operator

data OpUse v

Use of operator

Constructors

OpUse

Fields

opHalf :: HalfRelmap

Syntactic data of operator use

opSubmap :: [Relmap v]

Subrelmaps

Constructors

relmapSource

Arguments

:: OpUse v
Use of operator

-> String
Operator name

-> [String]
List of term names

-> Relmap v
Result relmap

Retrieve relation from dataset.

relmapConst

Arguments

:: OpUse v
Use of operator

-> String
Operator name

-> Rel v
Constant relation

-> Relmap v
Result relmap

Constant relmap.

relmapAlias

Arguments

:: OpUse v
Use of operator

-> Relmap v

-> Relmap v
Result relmap

Alias relmap.

relmapCalc

Arguments

:: OpUse v
Use of operator

-> String
Operator name

-> RelmapSub v
Calculation of operation

-> Relmap v
Result relmap

Make a non-confluent relmap.

relmapConfl

Arguments

:: OpUse v
Use of operator

-> String
Operator name

-> RelmapSub v
Calculation of operation

-> [Relmap v]
Subrelmaps

-> Relmap v
Result relmap

Make a confluent relmap.

Linker

relmapLink :: Map [Named (Relmap v)]

relmapLinker :: [Named (Relmap v)] -> Map (Relmap v)