Koshucode.Baala.Base.Syntax.Token

Description

Tokens in Koshucode

Token type

Word

Character sequence not including special characters, e.g., aa, r2d2, 12.0, etc.

Term name

Word beginning with slash, e.g., /aa. Term name like /a/b is used for nested relation, that means term /b in the relation of term /a.

Paren

Open and closed parens.

Comment

Text from ** to end of line is single line comment.

Space

Space characters.

Line

Source line.

data Token

Constructors

Word Int String
Text

TermN [String]
Term name

TermP [Int]
Term position

Open String
Open paren

Close String
Close paren

Space Int
N space characters

Comment String
Comment text

Line SourceLine
Source infomation

data SourceLine

Line number and content in source code

Constructors

SourceLine Int String [Token]

tokenTypeText :: Token -> String

Text of token type, one of "Word", "TermN", "TermP", "Open", "Close", "Space", "Comment", or "Line".

>>> tokenTypeText $ Word 0 "flower"
"Word"

Predicates

isBlankToken :: Token -> Bool

Test the token is blank, i.e., Comment, Line, or Space.

isLineToken :: Token -> Bool

Test the token is a line, i.e., Line.

isTermToken :: Token -> Bool

Test the token is a term, i.e., TermN or TermP.

Tokenizer

tokens :: String -> [Token]

Split string into list of tokens

>>> tokens "|-- R  /a A0 /b 31"
[Line 1 "|-- R  /a A0 /b 31",
 Word 0 "|--", Space 1, Word 0 "R", Space 2,
 TermN ["/a"], Space 1, Word 0 "A0", Space 1,
 TermN ["/b"], Space 1, Word 0 "31"]

untokens :: [Token] -> String

Convert back a token list to a source string.

untoken :: Token -> String

Othre functions

sweepToken :: Map [Token]

Remove blank tokens.

sweepLeft :: Map [Token]

Skip leading blank tokens.